Work on scout and worker moves. Also queen but she does not move

+++ include/Ants/Types/Worker.h
@@ -12,7 +12,7 @@ class Worker : public virtual Ant {
 private:
     bool has_food;
     bool major;
-    int pheromones_stock;
+    float pheromones_stock;
     int minor_lifespan;
 
 public:
